sandbox/tools/11_8_2025/requirements.txt
# Core JAX Ecosystem (Primary Framework)
jax>=0.4.0
jaxlib>=0.4.0
optax>=0.1.7
equinox>=0.11.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0
einops>=0.7.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Specialized Libraries for Consciousness Research
# ngc-learn  # Neural Generative Coding - uncomment when available

# Self-Organizing Maps
minisom>=2.3.0

# Probability and Statistics
numpyro>=0.13.0  # Probabilistic programming
distrax>=0.1.0   # DeepMind probability distributions (optional JAX extension)

# Time Series and Dynamics  
diffrax>=0.4.0   # Differential equation solvers for continuous dynamics

# Graph and Network Processing
networkx>=3.0
# jraph>=0.0.6   # JAX graph neural networks (optional)

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# GUI Framework (Japanese language support)
tkinter  # Standard library, but listed for clarity

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0  # Property-based testing

# Logging and Configuration
pyyaml>=6.0
toml>=0.10.0

# Optional: PyTorch for comparison/hybrid approaches
# torch>=2.0.0
# torchvision>=0.15.0

# Optional: Additional research libraries
# Brian2>=2.5.0      # Spiking neural networks (future extension)
# nest-simulator     # Large-scale neural simulations (future extension) 
# pymdp>=0.0.4       # Active inference (future extension)

# Documentation (development only)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Code Quality
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0

# Performance Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0